zenfone6: Remove the voip_rx source for the Earpiece route
Skype and TG seem use an diff audio route for voip and this device seem have an pb with mono and stereo route with the Voip.

With only mono mask, Skype working very well but TG cant change the vol up/down.
With stereo, TG working well but Skype/WhatsApp have lot of echoes and sound distortion.

Np in all cases with the Speaker route.

Change-Id: Icd1e8b76f7c7bd470f7a0942b2a4fd533c22e2db